api: Allow partial reprioritization
There were two issues:
1. A node would send all scheduled coflows in parallel.
This is fine if (as was assumed) lower-priority packets
are dropped by the host NIC. However, since this may not
be the case, we must explicitly send only the highest priority
coflow.

2. Reprioritization. Since we are now only sending one coflow at a time,
we can poll the scheduler to check if the priority for as-yet-unsent
coflows has been updated.